+++ packages/pwa-kit-react-sdk/src/utils/url.js
@@ -5,6 +5,8 @@
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
+import {getConfig} from '@salesforce/pwa-kit-runtime/utils/ssr-config'
+
 /**
  * Returns the application's origin.
  *
@@ -19,11 +21,16 @@
  * const url = `${getAppOrigin()}/path`
  */
 export const getAppOrigin = () => {
+    const config = getConfig()
     if (typeof window !== 'undefined') {
         return window.location.origin
     }
 
-    const {APP_ORIGIN} = process.env
+    const {APP_ORIGIN, X_FORWARDED_HOST} = process.env
+
+    if (config?.app?.useXForwardedHost && X_FORWARDED_HOST) {
+        return X_FORWARDED_HOST
+    }
 
     if (!APP_ORIGIN) {
         throw new Error(