+++ packages/pwa-kit-react-sdk/CHANGELOG.md
@@ -1,4 +1,6 @@
 ## v3.8.0-dev (Aug 08, 2024)
+- Extend getAppOrigin to use x-forwarded-host header when useXForwardedHost config flag is enabled [#2018](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2018)
+
 ## v3.7.0 (Aug 07, 2024)
 - Add `beforeHydrate` option to withReactQuery component [#1912](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/1912)
 - Add server side rendering performance metrics via query parameter `__server_timing` or environment variable `SERVER_TIMING`, the metrics is available in the console logs and response header `server-timing`. [#1895](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/1895)